server/requirements.txt
fastapi[standard]
pydantic-settings
joblib
scikit-learn==1.5.1
ipykernel
humps
pyjwt
passlib[bcrypt]
bcrypt==4.0.1
mlflow
fastapi-pagination
shap
Faker
scispacy
openpyxl
seaborn
matplotlib
numpy
imbalanced-learn